Developement of C++ modules

+++ src/include/math_aux.hh
@@ -25,7 +25,7 @@
 struct fit_poly {
     double mean_t, start_t, stop_t;
     double *mean_coords, *coeffs;
-    uint is_centered, deg;
+    size_t is_centered, deg;
 };
 
 #if 0
